backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
numpy
opencv-python
pandas
pytesseract
easyocr
faiss-cpu
sqlalchemy
mysql-connector-python
python-multipart
pydantic
pymysql
python-pptx
PyMuPDF
python-docx
cryptography
# google-generativeai (removed; switched to OpenAI)
comtypes; platform_system == "Windows"
pywin32; platform_system == "Windows"